Rename examples to artefacts.
