+++ yarn-project/acir-simulator/src/client/private_execution.test.ts
@@ -34,7 +34,7 @@ import {
   PendingCommitmentsContractAbi,
   TestContractAbi,
   ZkTokenContractAbi,
-} from '@aztec/noir-contracts/examples';
+} from '@aztec/noir-contracts/artifacts';
 import { PackedArguments, TxExecutionRequest } from '@aztec/types';
 
 import { jest } from '@jest/globals';
