+++ yarn-project/aztec-sandbox/src/examples/util.ts
@@ -1,6 +1,6 @@
 import { Contract, ContractDeployer, EthAddress, Wallet } from '@aztec/aztec.js';
 import { PortalERC20Abi, PortalERC20Bytecode, TokenPortalAbi, TokenPortalBytecode } from '@aztec/l1-artifacts';
-import { NonNativeTokenContractAbi } from '@aztec/noir-contracts/examples';
+import { NonNativeTokenContractAbi } from '@aztec/noir-contracts/artifacts';
 
 import type { Abi, Narrow } from 'abitype';
 import { Account, Chain, Hex, HttpTransport, PublicClient, WalletClient, getContract } from 'viem';
