+++ yarn-project/end-to-end/src/e2e_p2p_network.test.ts
@@ -10,7 +10,7 @@ import { CircuitsWasm, Point, getContractDeploymentInfo } from '@aztec/circuits.
 import { computeContractAddressFromPartial } from '@aztec/circuits.js/abis';
 import { Grumpkin } from '@aztec/circuits.js/barretenberg';
 import { DebugLogger } from '@aztec/foundation/log';
-import { TestContractAbi } from '@aztec/noir-contracts/examples';
+import { TestContractAbi } from '@aztec/noir-contracts/artifacts';
 import { BootstrapNode, P2PConfig, createLibP2PPeerId, exportLibP2PPeerIdToString } from '@aztec/p2p';
 import { TxStatus } from '@aztec/types';
