+++ yarn-project/noir-contracts/src/artifacts/child_contract.json
