+++ yarn-project/noir-contracts/src/artifacts/index.ts
