+++ yarn-project/noir-contracts/src/artifacts/parent_contract.json
