+++ yarn-project/noir-contracts/src/artifacts/public_private_contract.json
