+++ yarn-project/noir-contracts/src/artifacts/public_token_contract.json
