+++ yarn-project/noir-contracts/src/artifacts/rollup_native_asset_contract.json
