+++ yarn-project/noir-contracts/src/artifacts/test_contract.json
