+++ yarn-project/noir-contracts/src/artifacts/uniswap_contract.json
