+++ yarn-project/noir-contracts/src/scripts/copy_output.ts
@@ -108,21 +108,21 @@ const main = () => {
   const source = readFileSync(`${folder}/src/main.nr`).toString();
   const contractName = process.argv[3] ?? upperFirst(camelCase(name));
   const build = JSON.parse(readFileSync(`${folder}/target/main-${contractName}.json`).toString());
-  const examples = `src/examples`;
+  const artifacts = `src/artifacts`;
 
   const abi = {
     name: build.name,
     functions: getFunctions(source, build),
   };
 
-  const exampleFile = `${examples}/${snakeCase(name)}_contract.json`;
+  const exampleFile = `${artifacts}/${snakeCase(name)}_contract.json`;
   writeFileSync(exampleFile, JSON.stringify(abi, null, 2) + '\n');
   log(`Written ${exampleFile}`);
 
   writeToProject(abi);
 
   const typeFile = `src/types/${name}.ts`;
-  writeFileSync(typeFile, generateType(abi, '../examples/index.js'));
+  writeFileSync(typeFile, generateType(abi, '../artifacts/index.js'));
   log(`Written ${typeFile}`);
 };
