+++ yarn-project/noir-contracts/src/types/child.ts
@@ -13,7 +13,7 @@ import { ContractAbi } from '@aztec/foundation/abi';
 import { Fr, Point } from '@aztec/foundation/fields';
 import { AztecRPC } from '@aztec/types';
 
-import { ChildContractAbi } from '../examples/index.js';
+import { ChildContractAbi } from '../artifacts/index.js';
 
 /**
  * Type-safe interface for contract Child;
