+++ yarn-project/noir-contracts/src/types/non_native_token.ts
@@ -13,7 +13,7 @@ import { ContractAbi } from '@aztec/foundation/abi';
 import { Fr, Point } from '@aztec/foundation/fields';
 import { AztecRPC } from '@aztec/types';
 
-import { NonNativeTokenContractAbi } from '../examples/index.js';
+import { NonNativeTokenContractAbi } from '../artifacts/index.js';
 
 /**
  * Type-safe interface for contract NonNativeToken;
