+++ yarn-project/noir-contracts/src/types/pending_commitments.ts
@@ -13,7 +13,7 @@ import { ContractAbi } from '@aztec/foundation/abi';
 import { Fr, Point } from '@aztec/foundation/fields';
 import { AztecRPC } from '@aztec/types';
 
-import { PendingCommitmentsContractAbi } from '../examples/index.js';
+import { PendingCommitmentsContractAbi } from '../artifacts/index.js';
 
 /**
  * Type-safe interface for contract PendingCommitments;
