+++ yarn-project/noir-contracts/src/types/schnorr_account.ts
@@ -13,7 +13,7 @@ import { ContractAbi } from '@aztec/foundation/abi';
 import { Fr, Point } from '@aztec/foundation/fields';
 import { AztecRPC } from '@aztec/types';
 
-import { SchnorrAccountContractAbi } from '../examples/index.js';
+import { SchnorrAccountContractAbi } from '../artifacts/index.js';
 
 /**
  * Type-safe interface for contract SchnorrAccount;
