+++ yarn-project/noir-contracts/src/types/uniswap.ts
@@ -13,7 +13,7 @@ import { ContractAbi } from '@aztec/foundation/abi';
 import { Fr, Point } from '@aztec/foundation/fields';
 import { AztecRPC } from '@aztec/types';
 
-import { UniswapContractAbi } from '../examples/index.js';
+import { UniswapContractAbi } from '../artifacts/index.js';
 
 /**
  * Type-safe interface for contract Uniswap;
